requirements.txt
PySide6>=6.9.1
PySide6_Addons>=6.6.0
PySide6_Essentials>=6.6.0
adafruit-circuitpython-neopixel-spi==1.0.0
setuptools==75.5.0
